Allow all necessary tools to be sourced from Java toolchain (#29)
Certain Java build rules provided by this plugin accept and make use of
a Java toolchain outputted by `java_toolchain` in order to compile Java
code (via `javac`) and run test JARs (via `java`). Extend the rules so
that all JDK tools that they depend upon can be provided by a Java
toolchain; in practice the only additional tool that the toolchain needs
to provide is jlink (for `java_runtime_image`), although jlink was added
in Java 9 which means that `java_runtime_image` is only usable with a
`java_toolchain` that provides JDK 9 or newer. There's a run-time check
for this in the build rule.

This adds two new plugin configuration options, `JavaTool` and
`RunSelfExecutablesWithJavaTool`. The former is used whenever the `java`
binary is required and no toolchain is defined. The latter controls
whether this tool is invoked as part of the preamble in self-executable
JARs outputted by `java_binary`; if this is enabled, none of the build
commands executed by (or outputs from) the rules in this plugin require
a JRE or JDK to be present in the system path, although self-executable
JARs will need to be run with `java -jar` if they are moved outside of
the Please repo. This option causes breaking changes when enabled and is
therefore disabled by default.