`java_runtime_image`: rename `main_*` parameters and make them optional
The `main_module` and `main_class` parameters are problematic: their
naming disguises the fact that they're used in the `--launcher` argument
to jlink, and compels users to use them even though the `--launcher`
argument is optional when invoking jlink.

Rename the parameters to `launcher_module` and `launcher_class`
respectively (retaining aliases to the old names for backwards
compatibility), and make them optional - if omitted, the `--launcher`
argument will now be omitted when invoking jlink.

Fixes #38.