`java_library`: only stringify `javac_flags` when needed
It's difficult to reason about the type of `javac_flags` as the
`java_library` rule executes: it may be a `str`, `list` or `None`
depending on a number of plugin config options and the value passed into
`java_library` itself. Keep it as a `list` until it is interpolated into
the javac command, at which point stringify it.