`java_runtime_image`: exclude `java_toolchain` directory from JAR search (#49)
When a `java_toolchain` is used to build a `java_runtime_image`, the
toolchain's files are included in the Please sandbox when remote
execution is enabled, and should not be included in the `--module-path`
option to jlink. Exclude any JAR files that are part of the toolchain
from the jlink module path.

Fixes #47.